requirements.txt
```
streamlit
mediapipe
opencv-python
numpy
pillow
scikit-learn
scipy
```